+++ pfnet/cbranch.pyx
@@ -268,6 +268,17 @@ cdef class Branch:
 
         return cbranch.BRANCH_is_part_of_3_winding_transformer(self._c_ptr)
 
+    def is_zero_impedance(self):
+        """"
+        Determines whether branch is a zero-impedance branch.
+        
+        Returns
+        -------
+        flag : |TrueFalse|
+        """
+
+        return cbranch.BRANCH_is_zero_impedance(self._c_ptr)
+
     def has_y_correction(self):
         """
         Determines whether branch has y correction table.